+++ template/src/components/Organisms/TotalCharts/index.tsx
@@ -41,9 +41,9 @@ export default function TotalChart({ data }: { data?: [] }) {
             <Area
               type="monotone"
               dataKey="orders"
-              stroke={COLORS.primaryFF}
+              stroke={COLORS.primary}
               strokeWidth={2}
-              fill={COLORS.primaryFF}
+              fill={COLORS.primary}
               fillOpacity={0.09}
             />
           </AreaChart>
@@ -69,9 +69,9 @@ export default function TotalChart({ data }: { data?: [] }) {
             <Area
               type="monotone"
               dataKey="orders"
-              stroke={COLORS.primaryFF}
+              stroke={COLORS.primary}
               strokeWidth={2}
-              fill={COLORS.primaryFF}
+              fill={COLORS.primary}
               fillOpacity={0.09}
             />
           </AreaChart>
@@ -97,9 +97,9 @@ export default function TotalChart({ data }: { data?: [] }) {
             <Area
               type="monotone"
               dataKey="orders"
-              stroke={COLORS.primaryFF}
+              stroke={COLORS.primary}
               strokeWidth={2}
-              fill={COLORS.primaryFF}
+              fill={COLORS.primary}
               fillOpacity={0.09}
             />
           </AreaChart>