+++ template/src/components/Pages/Auth/MerchantInfo/index.tsx
@@ -1,7 +1,0 @@
-import MerchantInfoOrganism from "@/src/components/Organisms/MerchantInfoOrganism";
-
-function MerchantInfo() {
-  return <MerchantInfoOrganism />;
-}
-
-export default MerchantInfo;