+++ CHANGELOG.md
@@ -6,7 +6,8 @@ All notable changes to this project will be documented in this file.
 
 - Initial release of the Dashboard Vite Template.
 - Added React, Vite, TypeScript, Redux Toolkit, RTK Query, React Toastify, React-i18next, Day.js, Ant Design, React Hook Form, Atomic Design System, Husky, and ESLint.
+- New modern auth screens
+- Changed primary color
 
 ## [Upcoming]
 
-- New modern auth screens